♻️ refactor(command-palette): extract type declarations into types.ts
Move CommandStepChoice, StepChoicesResult, CommandStep, and PaletteCommand
from command-registry.ts into types.ts. Update all consumers to import
directly from types.ts while keeping re-exports in command-registry.ts
for backward compatibility.
