+++ browser-features/chrome/common/command-palette/command-registry.ts
@@ -22,58 +22,19 @@ import {
 } from "./multiInputCommand/index.ts";
 import { searchHistory, isHistoryCommand } from "./history-provider.ts";
 import { searchBookmarks, isBookmarkCommand } from "./bookmark-provider.ts";
+import type {
+  CommandStepChoice,
+  StepChoicesResult,
+  CommandStep,
+  PaletteCommand,
+} from "./types.ts";
 
-export interface CommandStepChoice {
-  /** Display label shown to the user */
-  label: string;
-  /** Actual value passed to the command function */
-  value: string;
-  /** Optional description shown below the label */
-  description?: string;
-}
-
-export interface StepChoicesResult {
-  choices: CommandStepChoice[];
-  /** Index of the choice to select by default when the step loads. */
-  defaultIndex?: number;
-  /** Whether more choices can be loaded via loadMore. */
-  hasMore?: boolean;
-  /** Load more choices and return the newly loaded choices with updated hasMore flag. */
-  loadMore?: () => Promise<{ choices: CommandStepChoice[]; hasMore: boolean }>;
-}
-
-export interface CommandStep {
-  id: string;
-  label: string;
-  placeholder: string;
-  /**
-   * Validate user input for this step.
-   * Return `true` to pass, or a string error message to display.
-   * Omit to skip validation.
-   */
-  validate?: (input: string) => boolean | string;
-  /**
-   * Static predefined choices for this step.
-   */
-  choices?: CommandStepChoice[];
-  /**
-   * Dynamic choices loader. Called when entering this step.
-   * If both `choices` and `choicesLoader` are defined, `choices` takes priority.
-   * Use this for choices that need to be fetched asynchronously (e.g., search engines).
-   */
-  choicesLoader?: () => Promise<CommandStepChoice[] | StepChoicesResult>;
-}
-
-export interface PaletteCommand {
-  id: string;
-  label: string;
-  description: string;
-  category: string;
-  keywords: string[];
-  fn: (win: Window, args?: Record<string, string>) => void;
-  /** If defined, the palette will prompt the user for each step before executing fn. */
-  steps?: CommandStep[];
-}
+export type {
+  CommandStepChoice,
+  StepChoicesResult,
+  CommandStep,
+  PaletteCommand,
+};
 
 const ACTION_CATEGORY_MAP: Record<string, string> = {
   "gecko-back": "navigation",
