+++ browser-features/chrome/common/command-palette/multiInputCommand/open-url.ts
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MPL-2.0
 
 import i18next from "i18next";
-import type { PaletteCommand } from "../command-registry.ts";
+import type { PaletteCommand } from "../types.ts";
 
 /**
  * Get hiragana reading keywords for a given action/command ID from i18n.
