+++ browser-features/chrome/common/command-palette/multiInputCommand/reopen-in-container.ts
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MPL-2.0
 
 import i18next from "i18next";
-import type { PaletteCommand, CommandStepChoice } from "../command-registry.ts";
+import type { PaletteCommand, CommandStepChoice } from "../types.ts";
 
 /**
  * Get hiragana reading keywords for a given action/command ID from i18n.
