+++ browser-features/chrome/common/command-palette/multiInputCommand/switcher/test/tab-switcher.test.ts
@@ -8,10 +8,8 @@ import {
   type TestCase,
 } from "../../../../../test/utils/test_harness.ts";
 import { tabSwitcherCommand, loadTabs } from "../tab-switcher.ts";
-import {
-  getPaletteCommands,
-  type PaletteCommand,
-} from "../../../command-registry.ts";
+import { getPaletteCommands } from "../../../command-registry.ts";
+import type { PaletteCommand } from "../../../types.ts";
 
 const rawTests: TestCase[] = [
   {
